backend/requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
python-multipart==0.0.9
pymongo==4.6.1
motor==3.3.2
python-keycloak==3.7.0
tavily-python==0.3.0
openai==1.12.0
ollama==0.1.6
python-dotenv==1.0.0
PyPDF2==3.0.1
docx2txt==0.8
langchain==0.1.5
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
pymilvus==2.3.3
sentence-transformers==2.2.2
numpy==1.24.3
tiktoken==0.5.2
